requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.26.0
google-genai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
